PlotExonResults(ExonModelStrain)	R Documentation

Plot Strain Interaction Plots for Exon Array Data

Description

Two functions for plotting Exon Array Data.

Usage


  PlotExonResults(resultsframe,ExpSet,dBPackage="mouseexonensembl.db")
  PlotExonResultsCI(resultsframe,ExpSet,dbPackage="mouseexonensembl.db")
  PlotExon(id, ExpSet, display=TRUE, savePlot=FALSE)
  PlotExon(id, ExpSet, display=FALSE, savePlot=TRUE)
  PlotExon(id, ExpSet)
  PlotExonCI(id, ExpSet, display=FALS, savePlot=TRUE)

Arguments

ExpSet 	 ExpressionSet containing probeset-level Exon Expression Data. Currently, only Mouse Exon 1.0 data is supported. The ExpressionSet phenoData should contain a column called 'Strain' where the two strains are coded 1 and 2.

id 	 Valid Ensembl Transcript or Gene ID.
Note a list of all valid Transcript or Gene IDs can be queried from the database package by using getAllTranscripts() or getAllGenes().

resultsframe 	 Data frame (produced by RunExonModelWorkflow) where there first column contains a list of valid Ensembl Transcript or Gene IDs.
Note a list of all valid Transcript or Gene IDs can be queried from the database package by using getAllTranscripts() or getAllGenes().

dbPackage 	 Name of database package containing Ensembl-Exon Array mapping (currently only mouseexonensembl.db exists).

savePlot 	 Boolean. If true, the PlotExon will save the plot in the current working directory. The name the function will output the name and location for the graphics file.

display 	 Boolean. If true, the PlotExon will save the plot in the current working directory. The name the function will output the name and location for the graphics file.

Details

PlotExon produces an interaction plot (in PNG format) for a given Ensembl Transcript/Gene ID. This plot may either be saved to disk in the current working directory. PlotExonCI produces a similar plot, but with error bars. The plots are named automatically using the following convention:

"gene symbol" + "ID" + "exon.plot.png"

PlotExonResults produces interaction plots for a data frame of results produced by RunExonModelWorkflow. PlotExonResultsCI is similar, but produces interaction plots with error bars. Each point in the plot consists of the mean expression of all probesets for a particular exon for that strain.

Note that both PlotExonResults and PlotExon only plots Exons which have probesets that exist in the current ExpressionSet, so that Exons that are not represented in the data are not shown. A list of Missing Exons can be found for a particular transcript/gene ID by using MissingExons.

See Also

RunExonModelWorkflow MissingExons

Examples


##not run
  #mouseexonensembl.db package required for example
  #load in sample dataset
  data(exontestdata)
  testTrans
  results <- RunExonModelWorkflow(TestSetTrans, testTrans, dBPackage="mouseexonensembl.db")
  #9 out of the 20 transcripts are multiple-exon transcripts
  results$multi
  return
  ##PlotExonResults(results$multi[1:5,], ExpSet=TestSetTrans)

[Package ExonModelStrain version 0.2 Index]